# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "vis-service API",
        "title": "vis-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/vis-service",
    "paths": {
...
        "/dashboard-data/metric": {
            "post": {
                "description": "从metric中获取数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DashboardData"
                ],
                "summary": "Metric",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "reqs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.QueryMetricsCombin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": true
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/dashboard-data/metric-range": {
            "post": {
                "description": "从metric中获取批量数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DashboardData"
                ],
                "summary": "Metric Range",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "reqs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.QueryMetricsRangeCombin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": true
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "entity.QueryMetrics": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "result_defs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ResultDef"
                    }
                }
            }
        },
        "entity.QueryMetricsCombin": {
            "type": "object",
            "properties": {
                "reqs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.QueryMetrics"
                    }
                }
            }
        },
        "entity.QueryMetricsRangeCombin": {
            "type": "object",
            "properties": {
                "end": {
                    "description": "格式 2006-01-02 15:04:05",
                    "type": "string"
                },
                "reqs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.QueryMetrics"
                    }
                },
                "start": {
                    "description": "格式 2006-01-02 15:04:05",
                    "type": "string"
                },
                "step": {
                    "description": "The maximum time between two slices within the boundaries.",
                    "type": "integer"
                }
            }
        },
        "entity.ResultDef": {
            "type": "object",
            "properties": {
                "decimal_places": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "description": "metric||value",
                    "type": "string"
                }
            }
        },
...
    }
}